Move copying helper functions to wagtail.core.models.copying
